requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
anthropic>=0.7.0
pyyaml>=6.0
click>=8.1.0
python-dateutil>=2.8.0
urllib3>=2.0.0
retrying>=1.3.0
tqdm>=4.65.0
pdfplumber>=0.10.0